[#100] Degration action is set via port builder
